Fixed link.sh so that backup files won't be overwritten with links.

+++ link.sh
@@ -13,7 +13,7 @@ fi
 FILES=".bashrc .vimrc .zshrc"
 for file in $FILES
 do
-    if [ -f ~/$file ]
+    if [ -f ~/$file -a ! -h ~/$file ]
     then
         mv -v ~/$file $BACKUP_DIR
     fi
